# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API de gestion commerciale (factures, bons de livraison, devis, articles, clients)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: Lister tous les documents (factures, devis, bons de livraison)
      parameters:
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/DocumentType"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/DocumentStatus"
        - name: clientId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentSummary"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Créer un nouveau document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    put:
      operationId: updateDocument
      tags: [documents]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
  /documents/{id}/status:
    patch:
      operationId: updateDocumentStatus
      tags: [documents]
      summary: Mettre à jour le statut (livré, payé, annulé...)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocumentStatusBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}/convert:
    post:
      operationId: convertDocument
      tags: [documents]
      summary: Convertir un document (devis→facture, facture→bon de livraison)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConvertDocumentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    DocumentType:
      type: string
      enum: [facture, devis, bon_livraison]
    DocumentStatus:
      type: string
      enum: [brouillon, valide, livre, paye, annule]
...
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        contactName:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        fiscalNumber:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, createdAt]
...
    DocumentLine:
      type: object
      properties:
        id:
          type: integer
        articleId:
          type: integer
          nullable: true
        reference:
          type: string
        designation:
          type: string
        quantite:
          type: number
          format: float
        unite:
          type: string
        prixUnitaire:
          type: number
          format: float
        remisePct:
          type: number
          format: float
        montantHt:
          type: number
          format: float
        depot:
          type: string
          nullable: true
      required: [id, reference, designation, quantite, unite, prixUnitaire, remisePct, montantHt]
    CreateDocumentLineBody:
      type: object
      properties:
        articleId:
          type: integer
          nullable: true
        reference:
          type: string
        designation:
          type: string
        quantite:
          type: number
          format: float
        unite:
          type: string
        prixUnitaire:
          type: number
          format: float
        remisePct:
          type: number
          format: float
        depot:
          type: string
          nullable: true
      required: [reference, designation, quantite, unite, prixUnitaire, remisePct]
    Document:
      type: object
      properties:
        id:
          type: integer
        type:
          $ref: "#/components/schemas/DocumentType"
        numero:
          type: string
        status:
          $ref: "#/components/schemas/DocumentStatus"
        date:
          type: string
          format: date
        echeance:
          type: string
          format: date
          nullable: true
        clientId:
          type: integer
        client:
          $ref: "#/components/schemas/Client"
        vendeur:
          type: string
          nullable: true
        reference:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        lines:
          type: array
          items:
            $ref: "#/components/schemas/DocumentLine"
        totalHt:
          type: number
          format: float
        totalRemise:
          type: number
          format: float
        totalTva:
          type: number
          format: float
        totalTtc:
          type: number
          format: float
        applyTva:
          type: boolean
        relatedDocumentId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        [id, type, numero, status, date, clientId, client, lines, totalHt, totalRemise, totalTva, totalTtc, applyTva, createdAt]
    DocumentSummary:
      type: object
      properties:
        id:
          type: integer
        type:
          $ref: "#/components/schemas/DocumentType"
        numero:
          type: string
        status:
          $ref: "#/components/schemas/DocumentStatus"
        date:
          type: string
          format: date
        echeance:
          type: string
          format: date
          nullable: true
        clientId:
          type: integer
        clientName:
          type: string
        totalTtc:
          type: number
          format: float
        createdAt:
          type: string
          format: date-time
      required: [id, type, numero, status, date, clientId, clientName, totalTtc, createdAt]
    CreateDocumentBody:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/DocumentType"
        date:
          type: string
          format: date
        echeance:
          type: string
          format: date
          nullable: true
        clientId:
          type: integer
        vendeur:
          type: string
          nullable: true
        reference:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        applyTva:
          type: boolean
        status:
          $ref: "#/components/schemas/DocumentStatus"
        lines:
          type: array
          items:
            $ref: "#/components/schemas/CreateDocumentLineBody"
      required: [type, date, clientId, applyTva, status, lines]
    UpdateDocumentStatusBody:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/DocumentStatus"
      required: [status]
    ConvertDocumentBody:
      type: object
      properties:
        targetType:
          $ref: "#/components/schemas/DocumentType"
      required: [targetType]
...