# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API de gestion commerciale (factures, bons de livraison, devis, articles, clients)
 * OpenAPI spec version: 0.1.0
 */
import type { Client } from "./client";
import type { DocumentLine } from "./documentLine";
import type { DocumentStatus } from "./documentStatus";
import type { DocumentType } from "./documentType";

export interface Document {
  id: number;
  type: DocumentType;
  numero: string;
  status: DocumentStatus;
  date: Date;
  echeance?: Date | null;
  clientId: number;
  client: Client;
  vendeur?: string | null;
  reference?: string | null;
  notes?: string | null;
  lines: DocumentLine[];
  totalHt: number;
  totalRemise: number;
  totalTva: number;
  totalTtc: number;
  applyTva: boolean;
  relatedDocumentId?: number | null;
  createdAt: Date;
}