# lib/api-zod/src/generated/types/documentLine.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API de gestion commerciale (factures, bons de livraison, devis, articles, clients)
 * OpenAPI spec version: 0.1.0
 */

export interface DocumentLine {
  id: number;
  articleId?: number | null;
  reference: string;
  designation: string;
  quantite: number;
  unite: string;
  prixUnitaire: number;
  remisePct: number;
  montantHt: number;
  depot?: string | null;
}